refactor: users module folder
