+++ packages/shared/src/users/user.types.ts
@@ -1,0 +1,5 @@
+import type { z } from "zod";
+import type { userSchema, userSummarySchema } from "./user.schema.js";
+
+export type UserSummary = z.infer<typeof userSummarySchema>;
+export type User = z.infer<typeof userSchema>;
